C++: move platform-related code to uPlatform/
This moves platform-related parts of the bundled C++ code into its own
output directory instead of interfering with the Uno/ directory, making
things a bit cleaner for version 3.0.
