+++ lib/Uno.Threading/Win32Helpers.uno
@@ -22,7 +22,7 @@ namespace Uno.Threading
         {
         }
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         [Require("Source.Include", "@{Uno.Exception:Include}")]
         public static MutexHandle CreateMutex()
         @{
@@ -34,25 +34,25 @@ namespace Uno.Threading
             return mutexHandle;
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static bool WaitOneMutex(MutexHandle mutexHandle, int millisecondsTimeout)
         @{
             return ::WaitForSingleObject(mutexHandle, millisecondsTimeout == -1 ? INFINITE : millisecondsTimeout) == WAIT_OBJECT_0;
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static void ReleaseMutex(MutexHandle mutexHandle)
         @{
             ::ReleaseMutex(mutexHandle);
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static void DisposeMutex(MutexHandle mutexHandle)
         @{
             ::CloseHandle(mutexHandle);
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static SemaphoreHandle CreateSemaphore(int initialCount, int maxCount)
         @{
             HANDLE semaphoreHandle = ::CreateSemaphoreW(nullptr, $0, $1, nullptr);
@@ -63,7 +63,7 @@ namespace Uno.Threading
             return semaphoreHandle;
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static bool WaitOneSemaphore(SemaphoreHandle semaphoreHandle, int timeoutMillis)
         @{
             if (timeoutMillis == -1)
@@ -72,7 +72,7 @@ namespace Uno.Threading
                 return ::WaitForSingleObject(semaphoreHandle, timeoutMillis) == WAIT_OBJECT_0;
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static int ReleaseSemaphore(SemaphoreHandle semaphoreHandle, int releaseCount)
         @{
             LONG res;
@@ -81,13 +81,13 @@ namespace Uno.Threading
             return res;
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static void DisposeSemaphore(SemaphoreHandle semaphoreHandle)
         @{
             ::CloseHandle(semaphoreHandle);
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static ResetEventHandle CreateResetEvent(bool initialState, bool autoReset)
         @{
             HANDLE resetEventHandle = ::CreateEvent(nullptr, autoReset, initialState, nullptr);
@@ -98,25 +98,25 @@ namespace Uno.Threading
             return resetEventHandle;
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static bool WaitOneResetEvent(ResetEventHandle resetEventHandle, int timeoutMillis)
         @{
             return ::WaitForSingleObject(resetEventHandle, timeoutMillis == -1 ? INFINITE : timeoutMillis) == WAIT_OBJECT_0;
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static bool ResetResetEvent(ResetEventHandle resetEventHandle)
         @{
             return ::ResetEvent(resetEventHandle) != FALSE;
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static bool SetResetEvent(ResetEventHandle resetEventHandle)
         @{
             return ::SetEvent(resetEventHandle) != FALSE;
         @}
 
-        [Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static void DisposeResetEvent(ResetEventHandle resetEventHandle)
         @{
             ::CloseHandle(resetEventHandle);
