+++ lib/UnoCore/UnoCore.unoproj
@@ -95,10 +95,6 @@
     "cpp/Uno/_config.h:File",
     "cpp/Uno/_internal.h:File",
     "cpp/Uno/_invoke.cpp:File",
-    "cpp/Uno/_main.cpp:File",
-    "cpp/Uno/_mainMobile.cpp:File",
-    "cpp/Uno/GLHelper.h:File",
-    "cpp/Uno/GraphicsContext.h:File",
     "cpp/Uno/Memory.cpp:File",
     "cpp/Uno/Memory.h:File",
     "cpp/Uno/NativeStackTrace.cpp:File",
@@ -111,7 +107,11 @@
     "cpp/Uno/Support.h:File",
     "cpp/Uno/SupportApple.mm:File",
     "cpp/Uno/Uno.h:File",
-    "cpp/Uno/WinAPIHelper.h:File",
+    "cpp/uPlatform/_main.cpp:File",
+    "cpp/uPlatform/_mainMobile.cpp:File",
+    "cpp/uPlatform/GLHelper.h:File",
+    "cpp/uPlatform/GraphicsContext.h:File",
+    "cpp/uPlatform/WinAPIHelper.h:File",
     "ios/@(Icons)/Contents.json:File",
     "ios/@(Project.Name).xcodeproj/project.pbxproj:File",
     "ios/@(Project.Name).xcodeproj/project.xcworkspace/contents.xcworkspacedata:File",
