+++ lib/UnoCore/cpp/Uno/Support.cpp
@@ -15,7 +15,7 @@ void uLogApple(const char* prefix, const char* format, va_list args);
 #endif
 
 #ifdef WIN32
-#include <Uno/WinAPIHelper.h>
+#include <uPlatform/WinAPIHelper.h>
 #else
 #include <pthread.h>
 #endif
