+++ lib/UnoCore/cpp/uPlatform/_mainMobile.cpp
