+++ lib/UnoCore/src/Uno/IO/FileStream.uno
@@ -8,7 +8,7 @@ namespace Uno.IO
     extern(CPLUSPLUS)
     struct FILEPtr
     {
-        [extern(WIN32) Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [extern(WIN32) Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static FILEPtr OpenOrThrow(string filename, string mode)
         {
             if (filename == null)
