+++ lib/UnoCore/src/Uno/IO/Path.uno
@@ -225,7 +225,7 @@ namespace Uno.IO
         }
 
         [extern(APPLE) Foreign(Language.ObjC)]
-        [extern(WIN32) Require("Source.Include", "Uno/WinAPIHelper.h")]
+        [extern(WIN32) Require("Source.Include", "uPlatform/WinAPIHelper.h")]
         public static string GetTempPath()
         {
             if defined(WIN32)
